Update auto-launch
Streamlined process for deleting instances made from the script being ran previously

+++ aws_auto_setup.sh
@@ -38,14 +38,29 @@ read -p $'What region would you like to deploy this in? (us-east-2, us-west-1, e
 export AWS_DEFAULT_REGION="$rgn"
 
 
-# Deletes any instances with browatch tag
-aws ec2 describe-instances --filters "Name=instance.group-name,Values='browatch'" --output text --query 'Reservations[*].Instances[*].InstanceId' > deleteid.txt
-doid=$(cat deleteid.txt)
+# Deletes any instances with browatch tag if they exist
+
+doid=$(aws ec2 describe-instances --filters "Name=instance.group-name,Values='browatch'" --output text --query 'Reservations[*].Instances[*].InstanceId')
+
+if [ -n "$doid" ]; then
 aws ec2 terminate-instances --instance-ids $doid
-
-echo "Waiting 30 seconds to allow any extra instances to terminate...(you may see an error if there were no instances to terminate)"
-
+echo "Waiting 30 seconds to allow any extra instances to terminate..."
 sleep 30
+
+else
+echo "No previous browatch instances found to terminate."
+
+fi
+
+
+
+#aws ec2 describe-instances --filters "Name=instance.group-name,Values='browatch'" --output text --query 'Reservations[*].Instances[*].InstanceId' > deleteid.txt 
+#doid=$(cat deleteid.txt)
+#aws ec2 terminate-instances --instance-ids $doid
+
+#echo "Waiting 30 seconds to allow any extra instances to terminate...(you may see an error if there were no instances to terminate)"
+
+#sleep 30
 
 echo "We will create your SSH keys now. They will be named browatch.pem. Any prior ones with the same name will be overwritten."
 sleep 5
